requirements.txt
transformers>=4.36.2
torch>=2.1.0
gradio>=4.0.0
pyttsx3>=2.90
huggingface_hub>=0.20.0
openai==0.28
python-dotenv